backend/requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
google-generativeai==0.8.5
google-genai==1.52.0
google-ai-generativelanguage==0.6.15
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-api-core==2.28.1
googleapis-common-protos==1.72.0
proto-plus==1.26.1
protobuf==5.29.5
grpcio==1.76.0
grpcio-status==1.71.2
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
urllib3==2.5.0
idna==3.11
charset-normalizer==3.4.4
sniffio==1.3.1
starlette==0.50.0
anyio==4.11.0
h11==0.16.0
pydantic==2.12.4
pydantic_core==2.41.5
typing_extensions==4.15.0
tenacity==9.1.2
tqdm==4.67.1
colorama==0.4.6
cachetools==6.2.2
httplib2==0.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.5
uritemplate==4.2.0
websockets==15.0.1
certifi==2025.11.12
